+++ python/sglang/srt/layers/quantization/fp8_utils.py
@@ -3,6 +3,13 @@ from typing import List, Optional, Tuple
 
 import torch
 
+try:
+    from vllm import _custom_ops as vllm_ops
+
+    VLLM_AVAILABLE = True
+except ImportError:
+    VLLM_AVAILABLE = False
+
 from sglang.srt.layers.quantization.fp8_kernel import (
     _enable_jit_deepgemm,
     per_token_group_quant_fp8,
@@ -17,30 +24,23 @@ from sglang.srt.utils import (
     is_hip,
 )
 
-try:
-    import vllm
-    from vllm import _custom_ops as ops
-
-    VLLM_AVAILABLE = True
-except ImportError:
-    VLLM_AVAILABLE = False
-
-use_vllm_cutlass_w8a8_fp8_kernel = get_bool_env_var("USE_VLLM_CUTLASS_W8A8_FP8_KERNEL")
-
 _is_hip = is_hip()
+_is_cuda = is_cuda()
+
 if _is_hip and get_bool_env_var("CK_MOE"):
     from aiter import gemm_a8w8_blockscale
 
-_is_cuda = is_cuda()
 if _is_cuda:
     from sgl_kernel import fp8_blockwise_scaled_mm, fp8_scaled_mm
 
     from sglang.srt.custom_op import scaled_fp8_quant as sgl_scaled_fp8_quant
     from sglang.srt.layers.quantization.fp8_kernel import sglang_per_token_quant_fp8
 
+use_vllm_cutlass_w8a8_fp8_kernel = get_bool_env_var("USE_VLLM_CUTLASS_W8A8_FP8_KERNEL")
+
 # Input scaling factors are no longer optional in _scaled_mm starting
 # from pytorch 2.5. Allocating a dummy tensor to pass as input_scale
-TORCH_DEVICE_IDENTITY = torch.ones(1, dtype=torch.float32)
+TORCH_DEVICE_IDENTITY = None
 
 _TORCH_VERSION = torch.__version__.split("+")[0]
 try:
@@ -264,7 +264,7 @@ def apply_fp8_linear(
             # final solution should be: 1. add support to per-tensor activation scaling.
             # 2. solve the torch.compile error from weight_scale.numel() == 1 and x_scale.numel() > 1 (below line#308)
             if _is_hip and weight_scale.numel() == 1:
-                qinput, x_scale = ops.scaled_fp8_quant(
+                qinput, x_scale = vllm_ops.scaled_fp8_quant(
                     input_2d,
                     input_scale,
                     use_per_token_if_dynamic=use_per_token_if_dynamic,
@@ -275,32 +275,29 @@ def apply_fp8_linear(
                 )
 
     if cutlass_fp8_supported:
-        try:
-            if VLLM_AVAILABLE and use_vllm_cutlass_w8a8_fp8_kernel:
-                # Fall back to vllm cutlass w8a8 fp8 kernel
-                output = ops.cutlass_scaled_mm(
-                    qinput,
-                    weight,
-                    out_dtype=input.dtype,
-                    scale_a=x_scale,
-                    scale_b=weight_scale,
-                    bias=bias,
-                )
-            else:
-                assert (
-                    weight_scale.numel() == weight.shape[1]
-                ), "cutlass w8a8 fp8 sgl-kernel only supports per-channel scale"
-                output = fp8_scaled_mm(
-                    qinput,
-                    weight,
-                    x_scale,
-                    weight_scale,
-                    out_dtype=input.dtype,
-                    bias=bias,
-                )
-            return output.view(*output_shape)
-        except (ImportError, NameError, AttributeError):
-            pass
+        if VLLM_AVAILABLE and use_vllm_cutlass_w8a8_fp8_kernel:
+            # Fall back to vllm cutlass w8a8 fp8 kernel
+            output = vllm_ops.cutlass_scaled_mm(
+                qinput,
+                weight,
+                out_dtype=input.dtype,
+                scale_a=x_scale,
+                scale_b=weight_scale,
+                bias=bias,
+            )
+        else:
+            assert (
+                weight_scale.numel() == weight.shape[1]
+            ), "cutlass w8a8 fp8 sgl-kernel only supports per-channel scale"
+            output = fp8_scaled_mm(
+                qinput,
+                weight,
+                x_scale,
+                weight_scale,
+                out_dtype=input.dtype,
+                bias=bias,
+            )
+        return output.view(*output_shape)
 
     # torch.scaled_mm supports per tensor weights + activations only
     # so fallback to naive if per channel or per token
@@ -343,8 +340,10 @@ def apply_fp8_linear(
 
             # Making sure the dummy tensor is on the same device as the weight
             global TORCH_DEVICE_IDENTITY
-            if TORCH_DEVICE_IDENTITY.device != weight.device:
-                TORCH_DEVICE_IDENTITY = TORCH_DEVICE_IDENTITY.to(weight.device)
+            if TORCH_DEVICE_IDENTITY is None:
+                TORCH_DEVICE_IDENTITY = torch.ones(
+                    1, dtype=torch.float32, device=weight.device
+                )
 
             # GEMM
             # This computes C = (X * W).
@@ -370,13 +369,6 @@ def apply_fp8_linear(
             if bias is not None:
                 output = output + bias
             return output.to(dtype=input.dtype).view(*output_shape)
-
-
-def maybe_create_device_identity():
-    # Allocate dummy ones tensor for torch._scaled_mm
-    global TORCH_DEVICE_IDENTITY
-    if TORCH_DEVICE_IDENTITY is None:
-        TORCH_DEVICE_IDENTITY = torch.ones(1, dtype=torch.float32)
 
 
 # Adapted from https://github.com/vllm-project/vllm/blob/main/vllm/model_executor/layers/quantization/utils/w8a8_utils.py
@@ -445,7 +437,7 @@ class Fp8LinearOp:
                     use_per_token_if_dynamic=use_per_token_if_dynamic,
                 )
             else:
-                qinput, x_scale = ops.scaled_fp8_quant(
+                qinput, x_scale = vllm_ops.scaled_fp8_quant(
                     input_2d,
                     input_scale,
                     scale_ub=input_scale_ub,
@@ -455,7 +447,7 @@ class Fp8LinearOp:
             # Fused GEMM_DQ
             if VLLM_AVAILABLE and use_vllm_cutlass_w8a8_fp8_kernel:
                 # Fall back to vllm cutlass w8a8 fp8 kernel
-                output = ops.cutlass_scaled_mm(
+                output = vllm_ops.cutlass_scaled_mm(
                     qinput,
                     weight,
                     out_dtype=input.dtype,
@@ -489,7 +481,7 @@ class Fp8LinearOp:
                     use_per_token_if_dynamic=use_per_token_if_dynamic,
                 )
             else:
-                qinput, x_scale = ops.scaled_fp8_quant(
+                qinput, x_scale = vllm_ops.scaled_fp8_quant(
                     input_2d,
                     input_scale,
                     num_token_padding=self.output_padding,
