+++ python/sglang/srt/reasoning_parser.py
@@ -1,4 +1,3 @@
-import re
 from typing import Dict, Tuple
